+++ src/packages/eslint-config-kiwicom/changeNextVersionErrorLevel.js
@@ -2,12 +2,16 @@
 
 /*::
 
-type EslintConfigErrorLevel = 0 | 1 | 2 | 'off' | 'warn' | 'error' | 3; // "3" is our speciality
+// Level "3" is our NEXT_VERSION_ERROR speciality:
+type EslintConfigErrorLevel = 0 | 1 | 2 | 'off' | 'warn' | 'error' | 3;
+
+type EslintConfigValues =
+  | EslintConfigErrorLevel
+  | [EslintConfigErrorLevel, any]
+  | [EslintConfigErrorLevel, any, any];
+
 type EslintConfig = {
-  [string]:
-    | EslintConfigErrorLevel
-    | [EslintConfigErrorLevel, any]
-    | [EslintConfigErrorLevel, any, any],
+  [string]: EslintConfigValues,
 };
 
 */
@@ -15,20 +19,16 @@ type EslintConfig = {
 module.exports = function changeNextVersionErrorLevel(
   config /*: EslintConfig */,
   level /*: 0|1|2 */,
-) {
+) /*: EslintConfig */ {
   return Object.entries(config).reduce((acc, [ruleName, value]) => {
-    let newValue = -1;
-
-    if (value === 3) {
-      newValue = level;
-    } else if (Array.isArray(value)) {
-      newValue = value;
-      newValue[0] = level;
+    if (Array.isArray(value) && value[0] === 3) {
+      const [, ...config] = value;
+      acc[ruleName] = [level, ...config];
+    } else if (value === 3) {
+      acc[ruleName] = level;
     } else {
-      newValue = value;
+      acc[ruleName] = value;
     }
-
-    acc[ruleName] = newValue;
     return acc;
   }, {});
 };